Remove postgres container from production docker-compose.yml
This aims to reflect more likely situation in production deployments
where a remote postgres database is used with relevant backup schedule
and other disaster recovery improvements.

In case it is desirable to run postgres in the same docker compose
service, it is still possible to add and link it as usual.