Add boi2024 specific HTTPS configuration.
This uses docker-nginx-certbot and requires .env.nginx to be configured
with operator's contact email and FQDN for the translate application.

This is kept separate from IOI nginx configuration primarily because
there is no straightforward way to glue this all together in HTTP-only
way. IOI configuration (virtually identical!) then can be used for debug
purposes.