Make SC name/acronym configurable.
This is required for BOI because instead of ISC we have HSC (host
scientific committee).

The implementation uses SC_TITLE setting to divine the privileged
username for task ownership. This implies that administrator must create
user with username matching SC_TITLE configuration option for
translation workflow to work.

This also highlights that this probably should not be a
hardcoded/configurable thing, but an acl group (similar to editors)
with separate localization option (e.g. separate country).